Allow TCP Server to reuse address after restart
After killing (Ctrl-C) a rosbridge_tcp server instance which has
connected clients, starting a new instance (on the same port) can
fail with the error: '[Errno 98] Address already in use'. Although the
node retries until the server starts, this can take up to a few minutes.

Instruct the ThreadingTCPServer to allow the reuse of the same address.