add graceful_shutdown() method to advertise_service capability
This gives the service a bit of time to cancel any in-flight service requests (which should fix #265).
This is important because we busy-wait for a rosbridge response for service calls and those threads do not get stopped otherwise.
Also, rospy service clients do not currently support timeouts, so any clients would be stuck too.

A new test case in test_service_capabilities.py verifies the fix works