Add cbor-raw compression
The CBOR compression is already a huge win over JSON or PNG encoding,
but it’s still suboptimal in some situations. This PR adds support for
getting messages in their raw binary (ROS-serialized) format. This has
benefits in the following cases:
- Your application already knows how to parse messages in bag files
  (e.g. using [rosbag.js](https://github.com/cruise-automation/rosbag.js),
  which means that now you can use consistent code paths for both bags
  and live messages.
- You want to parse messages as late as possible, or in parallel, e.g.
  only in the thread or WebWorker that cares about the message. Delaying
  the parsing of the message means that moving or copying the message to
  the thread is cheaper when its in binary form, since no serialization
  between threads is necessary.
- You only care about part of the message, and don't need to parse the
  rest of it.
- You really care about performance; no conversion between the ROS
  binary format and CBOR is done in the rosbridge_sever.

For [Webviz](https://webviz.io) we care about all of these cases, and so
we’d really like to be able to use raw binary ROS messages.

This seemed easy to support by extending the current CBOR-format. We
simply encode the `outgoing_msg` (which contains stuff like the
operation and the topic name) with CBOR, just like with the existing
CBOR encoding. The only difference is that in the current CBOR encoding,
the “msg” field is itself a CBOR structure that encodes the message,
where in our encoding it’s a CBOR object that contains a “bytes” field,
which is just a bytearray with the raw binary message.

Besides the “bytes” fields we also added “secs” and “nsecs” fields with
the current time, which is useful when using `use_sim_time` set to true.
Webviz needs this to properly show the simulated received time of
messages.

Now that we have the raw messages, we need to be able to parse them on
the client. Bags contain the raw message definitions as stored on the
`._full_text` field on message instances, so we added a service call for
accessing these raw message definitions for all topics in one call.

With both of these changes combined, you can now easily parse messages
as late, parallel, partially, and fast as possible. In Webviz we intend
to use this by either using [rosbag.js](https://github.com/cruise-automation/rosbag.js)
or by parsing using the ROS deserializer compiled to WebAssembly, and by
doing this in the WebWorker that actually controls the 3d canvas using
the OffscreenCanvas API. This way we can have multiple rendering
surfaces running from parallel threads. We already do this for some
panels in Webviz, such as the Image panel, and we’ve seen substantial
speedups with this method.

We’ll post a companion PR to roslibjs.

Documentation: we updated the ROSBRIDGE_PROTOCOL.md file with the new
changes.

Test plan: we added a new test for the WebSocket server, modeled after
the existing test. We also tested this end-to-end with our Webviz
client, and it seems to work great.