Add support for the "id" field in subscribe. For the "subscribe" and "unsubscribe" operations, the "id" field is now taken into account. This means that when it is provided, the subsequent "publish" operations, from the bridge to the client, will carry the "id" field with the same value.
Also, if multiple subscriptions are done on the same topic with
different values for "id", the related messages are sent as many times
as there are subscriptions with different "id" values.

If one uses the "unsubscribe" operation on a topic without the "id"
field, all subscriptions to that topic are closed.