Explicitly rename param subnode (#707)
**Public API Changes**
None(ish)

The param subnode is now named `{parent_node_name}_params` when the rosapi node is renamed. Previously it would have the same name.


**Description**
Previously, without any renaming, running the `rosapi` node would spawn a `rosapi_params` subnode. If the top-level node was renamed (or had a name set by a launch file) the param subnode ended up with the same name as the top-level node. This fixes the behaviour by explicitly setting the name of the subnode to the same as the top-level node with `_params` as a suffix.

Unlike the solution described in [the issue](https://github.com/RobotWebTools/rosbridge_suite/issues/704), this preserves any other global arguments. This means when changing namespace of the top-level node, the subnode will have a matching namespace.

Fixes #704