Non standard msg modules (#735)
This is a bug fix relating to parsing the python module and class names of message and service type strings when non-standard submodule naming is used.

I have an application that has generated messages in a namespace that is not msg. For example, I want to subscribe to a message type 'my_package_interface/msg_core/MyMessage'. However, the ros_loader.py method _get_msg_class discards submodule information 'msg_core' and assumes the subname will be named 'msg' causing an InvalidModuleException to be raised.

I like that the message type parsing is flexible and able to accommodate ROS1 style types that don't have a submodule subname in the type string. However, I think that logic is too strict because it assumes the submodule subname and prevents any naming other than msg.

Note that I am trying to use the term "submodule" as it is used in https://docs.python.org/3/tutorial/modules.html even though it could be confused with git submodule.

Co-authored-by: William Wedler <william.wedler@resquared.com>