Fixed /get_param service for arrays (#776)
**Public API Changes**
None


**Description**
The missing comma resulted in a wrong mapping of array parameters. For example, requesting a string array only returned an empty value because the type was given as index 9 (entry 10) but the type_mapping only had 9 entries. 

The error should be obvious, so I have not included an extra test.