Minor performance improvements (#809)
* Remove write lock.

The function `IOLoop.add_callback` is thread-safe and doesn't need to
be protected for mutual access. And since the `self.prewrite_mssage` is
only executed in the tornado event loop, there is no need for the write
lock.

* Migrate from decorated coroutine to native coroutine.

Native coroutines are generally faster as mentioned in the tornado docs:
https://www.tornadoweb.org/en/stable/guide/coroutines.html#native-vs-decorated-coroutines

* Use executor.spin_once instead of rclpy.spin_once.

This is more explicit and avoids some computationaly overhead.

* Use one callback group per subscriber.

While this does not have any effect on the current single-threaded
implementation, it will pave the way for a multi-threaded executor
later on.