fix list field bug
The issue was with the array's rostype being in a format like <float, 1>, which includes a number after the type separated by a comma. When this is passed to _from_list_inst(), it returns a rostype like float, 1, which does not qualify as a primitive type in the _from_list_inst() and _from_inst() methods.

As a solution, I added a patch in the _from_list_inst() method that treats the string before the comma as the rostype.